[tests] Simpler color output
From the comment in the Gemfile, it seems that we only want
`minitest-reporters` for output with colors. `minitest/pride` is builtin
in minitest and much simpler. I had no special opinion on removing this
but I was trying to debug some problems with the test setup, so I
started "removing things" to isolate those. After fixing the problems,
this seems like a good move.