[CI] Make sure latest bundler is used in travis
It seems like there's is a new infrastructure where an old version of
bundler is provided by default. We use newer features such as the
ability of passing a block to `source` in the Gemfile, so we need to
make sure an up to date version is used.

My guess is that [the
fix](https://github.com/travis-ci/travis-build/issues/675) to
travis-ci/travis-ci#5698 broke this.