[CSS] Actually break long words
In 622ecac and 50fa084 we fixed an issue were words would be broken at
random places. But with that fix, we broke a previous fix for #100
(c9bf121 and bef91f8) because we don't break long words at all. This is
a compromise between the two: break just words when supported, fallback
to break anywhere otherwise (apparently this is just Safari and iOS).