[tests] Fix random test failure
When running tests with seed 34007, we would get the following failure
in "AuthenticatedAdListing#test_0003_lists reserved ads in users
location in home page":

```
ActionController::RoutingError: No route matches [GET] "/admin/jobs"
```

This is because some tests wouldn't reset the session after themselves.
To fix it, reset the session in a global after hook.