[gems] Get rid of `blueprint-rails` gem dependency
The gem hasn't been updated for 3 years, and there are now better
alternatives. So I get rid of the dependency and move any CSS reset
styles to a separate files. All common global styles are in a single
place now, not split between the blueprint gem and the common.css file.
This should make migrating to the new design easier.