[users] Remove visual inconsistencies in tabs
In our efforts to remove links to the current tab, tab styles get a
little bit messed up. I think doing this is not worth the effort and
we're not doing it anywhere else (home page, status filters, etc). Plus,
I haven't seen this practice in a lot of places out there and I'm not
sure it's necessarily positive.