Fix incorrect instance variable name
This error was causing "Ver anuncios en tu ciudad o pueblo" to always be
displayed instead of suggesting the city guessed from the IP when
possible. Even if the IP geolocation might not be the best, this is
still better than the current situation since the old functionality is
always provided via the "Cambiar ciudad" link.