[data] Add a task to migrate duplicate users
There are 1111 duplicated users in database. This is because we used to
allow insertion of mixed-case emails in database. So now if
"EMAIL@example.com" is already in database, insertion of a user with
email "email@example.com" won't fail, because what `devise` checks is
that the downcased version of the email to be inserted is not already in
database.

For the migration, I use several criteria to choose the preferred user.
In order of precedence:
- Last login time.
- Last published ad.
- Last confirmation.
- Last sent confirmation email.

After that, all data is updated to point to the preferred user and the
duplicated users are deleted.