Migrator to remove all invalid woeids in DB
We currently have no ads with invalid woeid, but have a lot of users
with those. This commit intends to remove those from the DB in a sane
way. The migration has two phases:

* First, invalid woeids are detecting by querying the Yahoo API for the
type of each woeid and checking that it's a town.

* Then, an alternative "town woeid" is found for every invalid woeid.
Here, the criteria, in order from most to less priority is:
  - If woeid refers to a postal code, district or any other entity of
    lower degree than a town, use the containing town.
  - If the user with the invalid woeid has published ads, use woeid of
    the last published ad.
  - If woeid refers to a region, country or any other entity of higher
    degree than a town, use any contained town.
  - If woeid is invalid, use Madrid.