Fix hreflang errors in google webmastertools
The problem is that the pages where `page` is specified as a query
string parameter have no "return tag", because the alternative URLs
have no query string params, so they can't point back to the page.

The solution is to always use URLs using "page" as a url segment. The
old URLs of the form `https://nolotiro.org/es/ad/listuser/id/1?page=1`,
but all URLs generated internally have the new format.

In addition, we do the same for the `status` and `type` parameters.