Normalize users's primary key type
It was currently an unsigned int but this runs into trouble when trying
to create foreign key references since the types don't match and mysql
is very picky with foreign key types.

This is probably not compatible with postgreSQL but current migrations
don't even run from scratch, so we'll probably won't use them if
migrating to postgreSQL. If fact the new state is probably more
compatible with postgreSQL than it was, although postgreSQL does not
seem as picky as MySQL regarding data types.