Better API mocking
Partially revert b4802ac in favor of a monkeypatching, but cleaner,
solution.

Now tests pass without an internet connection and they take only
~23secs in my machine (33% faster). I still keep the tests running
against the real API in CI for peace of mind.

We changed the default city for tests to "Rio de Janeiro" from "Madrid",
because it has no cities with similar names, so Yahoo response was
easier to mock.