Implement multi-step "edge case" registration
When the omniauth hash comes without email, we need to implement another
step so that users can verify their email. This also includes dropping
the email as the authentication key for the omniauth case, since it's
not guaranteed to be there. When need to look in the identities table
where we store the omniauth credentials.