Stop forbidding short usernames
This fixes an error that prevents some legacy users to login. These
users already have a username shorter than 3 characters. We they login
and the system wants to upgrade them to the new password system, saving
the record fails because their username is too short.

Just drop this restriction from the model since there's no good reason
to enforce it and it's not even enforced in DB.