Remove references to Río de Janeiro in tests
They only reason we introduced it is because it didn't have cities with
similar names so it allowed us to touch the Yahoo API less. Now that
we're properly mocking the API responses, we no longer need it.

This also fixes a bug where the responses of the Yahoo search for
similar city names were recorded under the woeid code and not the search
key being used.