Add a task to remove username duplications
The idea here is to find all duplicated usernames (case insensitively),
and rename them to eliminate the duplications. We just rename all the
users involved in the conflict to the old username plus a sequence
number. On a later commit, we'll enable changing the username in the
profile. The first one to use this feature will be able to recover its
old username :)