Fix some errors with google adsense loading
We're already filling all banners at the bottom of the page, so that's
too many calls to `adsbygoogle.push()`.

Fixes the following javascript error

```
TagError: adsbygoogle.push() error: All ins elements in the DOM with
class=adsbygoogle already have ads in them.
```

Note that in the other calls we're not passing the google cliente code
nor the flag to enable page ads. None of them are necessary since the
client code is also set through "data" attributes and we're not using
page ads.