Extract Yahoo API info to a class
This is not only a refactoring but a big performance improvement since
we no longer touch the API again for each search result. The information
is already provided! For example, in the "San Pedro" case, the search
would return 452 results and we would subsequently call the Yahoo API
again 452 times to resolve each name. This is not necessary.