Fix bug where ad data currently in DB was invalid
Because we're overriding body & title to escape privacy data, if we
validate ads currently in DB, some entries would be invalid because
checks are made against the filtered text, and sometimes that text
wouldn't pass length validations that the full text passes.

The solution is to not override those, but define separate methods.

I also reduced the size of the sample texts, so that they would catch
the bug is this regressed somehow.