Align comment form with comments
To do this, we move the form inside the comment list div, and make the
wrapper div styles for the form extend the styles from the comment
wrapper div.

At that point, the comments/show template is no longer necessary and
since it's a bit confusing that a "show" template is used to render a
list of resources, I just removed it.