Remove similar cities "feature"
There's several reasons to remove this:

* Not necessary. When choosing to change the location, the user already
knows the full name of it and the number of ads. Besides, current
selector already offers cities with similar names to the one typed. Why
would she want a location with a similar name after choosing? In the
case of an actual typo + erroneous selection, she can just change
location again. In the future, we should offer suggestions on the fly
through js to make this better.

* Behaves bad with bots & Yahoo API. The list of suggestions is huge for
some places with very standard names. Because of this, bots keep
"discovering new places" and querying the Yahoo API until limits are
reached and errors are thrown.