Allow self conversations
This is very unlikely to cause any problems, apart for decreasing the
sociability of some users... :) Popular messaging apps like Telegram
allow it and we're trying to be more like them. But the main reason is
that I'm trying to simplify all this.

Since we have no redirect inside the method we can remove the
`performed?` check too.