Manage conversations ourselves
With `send_message` and `reply_to_conversation` we have no control
whatsoever over the conversations and messages being created. Since we
want to authorize those to implement blockings, I decided to get rid of
those and implement the same thing in the controller.

This is also good because at some point in the near future, we might not
only have conversations as message aggregators, but also things like
"candidacies" and/or "petitions". This should make it easier to add
those.