Completely RESTfull conversations
We change once more how messaging is handled by saving the parent
conversation first and then the associated objects.

Now every controller action creates/updates a conversation object, which
is then validated and only if successfully saved email notification is
sent.

This has several benefits:

* When we add authorizations we'll always have an actual conversation
object which we can authorize. The `envelope_for` method takes care of
creating the necessary associations so that validations and
authorizations can be properly run.

* Since we run save on the parent object, we don't need special flags
on the `belongs_to` association to "save backwards" or special
mechanisms (the reply parameter) to touch the conversation on update.

* We prevent running validations twice.

* We don't need the virtual `recipient` attribute to temporarily store
the recipient, since it's properly stored through associations. Thus, we
don't need to reset it either.