Remove friendships controller test
At some point I want to have only integration tests, specially once we
upgrade to Rails 5 because integration tests have apparently gotten much
faster. In this particular case, I don't see any situations that
wouldn't be catch by the corresponding integration test.

Plus, the current assertions are very weak. As an example, the test that
was supposed to be testing the `destroy` action was actually calling the
`create` action and it was passing anyways.