Conversation should be touched on reply
This was properly working in mailboxer because the `belongs_to` had a
`touch: true` flag. However, I don't like this kind of automatic
behavior. As a matter of fact, I started this patch by restoring that
flag but my test would be messed up by it because, while I explicitly
set the `updated_at` flag of the conversation in the test, it was
silently getting overriden by the flag.

Let's be explicit.