Get rid of `thinking_sphinx`
It gets in the middle and it doesn't play nice with ActiveRecord scopes.
Besides, Nolotiro has a short ad life cycle, and currently few ads, so a
complex search is not really necessary.

This will allow easier merging of `SearchController`, `AdsController`
and `WoeidController` as it's planned.

Finally, the search as it was previously implemented was a little bit
confusing because it would search in the ad body but the ad body is not
fully displayed in the ads list.