Change conversation cleanup task
We never run the old conversation cleanup task and now we've switched
back to the "old way": single receipt conversations are allowed. So we
add a task to remove the "orphan receipts" that have appeared during the
time where the `dependent: destroy` option in the `receipts` user
association was not there.