Stop linking directly to localeapp
We've literally never got anything reported there. I think this is
because the technical interface is unfriendly to non-technical people.
Let's direct them to the regular contact form and once we start a
conversation, introduce them to the translation tool. Besides, direct
contact is better for reporting. I'm not sure we know how to just
"report a bad translation" in localeapp (if you can even do that), let
alone a nolotiro.org user.