Fix `meta_title` tests
Changes SEO a bit. We stop dumping "regalo segundamano gratis" to
every ad's meta title. This is not always accurate, not every ad is a
present and not every ad is second hand. Regarding "gratis", I think
there's too much SEO competition on that word, I would leave it alone.