Normalize timestamps all around the app
Show "humanized timestamps" in a consistent way everywhere. Just use
what Rails offers instead of ~10_000 lines of javascript.

I've heard the "move processing to the client side" argument but I don't
think this is usually the bottleneck and I don't think this kind of
optimizations are something we want to worry about at this stage.