Cleanup old migrations in DB
* Old migrations don't really work. The first one is adding a
`deleted_at` column to the `ads` table, which is not created yet at that
time.

* They hard code `ActiveRecord` models so they get broken every time we
rename a model, for example.

* Keeping them in version control adds complexity. For example, it makes
grepping harder, makes the repo heavier, etc.

So what I did is to rollback until I got an error, and add a migration
to prune all migrations older than the latest successfully rolledback
migration. At a later stage this migration (already marked as run when
deployed) will be simply substituted by an initial DB schema, so
migrations can be run (in development or test, obviously), from scratch.