Stop checking SPAM per ad
We already have the "banning" functionality, and our spammers don't
"become" spammers, they're spammers from the beginning and all of their
content is SPAM.  So it's enough to flag the user, instead of every ad.

This has the nice side effect of automatically kicking out the spammer
as soon as SPAM is detected.