Remove a test
This is not the behavior I want to "tend to". My idea is that we should
do all moderation from the admin console and we should see the exact
same site as the rest of the users.

I'm still not removing the feature of deleting ads from the website
because my colleagues might complain... but I'll start by removing the
test case while I work on improving the admin console so that nobody
misses this.