Unique meaningful page title for listings
We were showing duplicated titles for a lot of listings. We
were advertising all of them as listings of presents while some of them
are not. We were always dropping the word "segundamano" there which was
giving us a good rank for "segundamano", but maybe losing potential for
more accurate searches that would provide us with a higher click rate.
Also, we were always using the word "gratis". This word is implied in
the word present, so we can skip it.

In general, I'm trying to be accurate & honest representing the content
of the page.