Improve mobile usability
Right now it's very hard to navigate nolotiro from mobile devices. Links
are very small and very close to each other. Google agrees with this in
its:

* "Size tap targets appropriately" warning in PageSpeed Insights
* "Clickable elements too close together" in Webmaster tools' Mobile
Accessibility.

We fix this by making the whole entry in an ad listing a single link to
the ad. Once we are inside the Ad, we also provide links to the city and
the user's profile.

One might argue that the site also loses usability with this change
since we remove direct links to the ad city and ad owners profile. I
find those direct link very rarely used and their still available from
the ad's page, so I don't think this is an issue.