Fix bad canonicals in homepage
Just using `url_for` was "kind of" doing what we wanted: the home page &
the listall page are actually different URLs for the same content, so we
need to use the same canonical for both. That was working.

But we were using the full "ugly" URL as the canonical
"ad/listall/ad_type/give" instead of the home page itself. This commit
fixes that.

I also added regression tests for this since this is rather critical
stuff for SEO and I don't want to break it or unintentionally change it
no more.