Fix alternates being too smart
They were previously doing some canonization due to how `url_for` works.
What we want is to just replace the locale with all the available
alternative locales. So just do that.

I also added regression tests for this since this is rather critical
stuff for SEO and I don't want to break it or unintentionally change it
no more.