Restore padding in image container
It was incorrect to remove it in 9484a75. The problem was the fixed
width of the container, not the padding. Actually I decided to make it a
little big bigger. With the old value, the text still gets too closed to
the picture.

Note to self, add images to staging copies so we can see these visual
problems behore hand.