Better filtering of sensitive information
Previously the content was just removed, now it's masked with a "PRIVATE
INFORMATION HIDDEN" message. I think this is less surprising for users
since the content does not just "dissapear". It also helps explaining a
bit why we do what we do.

It also simplifies the code since we can remove the length validation on
the filtered content that was originally introduced to prevent empty
titles from sneaking into the web. Previously, if an ad was published
with a "example@example.org" title, it would show as empty unless we
validated the filtered content. Now it doesn't matter because we are
using an alternative text.

In the future, we might want to add a link to the privacy policy to this
informational text, but this is an improvement over what we have.